+++ Trarizon.Library.GeneratorToolkit/CustomSymbolDisplayFormats.cs
@@ -1,0 +1,7 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace Trarizon.Library.GeneratorToolkit;
+public static class CustomSymbolDisplayFormats
+{
+    public static readonly SymbolDisplayFormat DefaultWithoutGeneric = SymbolDisplayFormat.CSharpErrorMessageFormat.WithGenericsOptions(SymbolDisplayGenericsOptions.None);
+}