+++ Trarizon.Library.GeneratorToolkit/_ref/Trarizon.Library/Collections/Extensions/EnumerableQueries/Helpers/.cs
@@ -1,4 +1,0 @@
-﻿namespace Trarizon.Library.Collections.Extensions;
-internal static partial class EnumerableQuery
-{
-}