+++ Trarizon.Library.SourceGenerator/DiagnosticIds.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Trarizon.Library.SourceGenerator;
+internal static class DiagnosticIds
+{
+    public const string FriendAnalyzer = "FRI";
+}