+++ Trarizon.Library/CodeAnalysis/FriendAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System.Diagnostics;
+
+namespace Trarizon.Library.CodeAnalysis;
+[AttributeUsage(
+    AttributeTargets.Constructor |
+    AttributeTargets.Method |
+    AttributeTargets.Property |
+    AttributeTargets.Field |
+    AttributeTargets.Event)]
+[Conditional("CODE_ANALYSIS")]
+public sealed class FriendAttribute(params Type[] friendTypes) : Attribute
+{
+    public Type[] FriendTypes { get; } = friendTypes;
+}